requirements.txt
python-telegram-bot==20.6
gradio==4.26.0
supabase==1.0.3
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.29.0
requests==2.31.0
pydantic==2.7.1
